inquiry-service/requirements.txt
fastapi
uvicorn
boto3
pydantic
python-jose